Footer bootstrap layout update.
Containers now in usage; needed separate col containers for light and dark grey areas of footer so that the colors could be maintained at 100% page width.

Height of containers needs to be adjusted. Links need further separation. .svg height within container needs to be adjusted. As well, content within containers should not be at the exact top of the containers in question.